src/vis_receive/requirements.txt
--index-url https://nexus.engageska-portugal.pt/repository/pypi/simple
--extra-index-url https://pypi.org/simple
ska-logging
ska-sdp-workflow